Create server.c
For rapid testing I added this TCP server that uses winsock library, run this on the host windows machine and the keylogger on a Windows VM or physical, whatever you want. Make sure to be on the same network, if you can't read target data then check the firewall to allow network traffic in between.